Rename classes as result of descussion on #594
