+++ example/src/main/java/io/netty/example/proxy/HexDumpProxyFrontendHandler.java
@@ -15,7 +15,7 @@
  */
 package io.netty.example.proxy;
 
-import io.netty.bootstrap.ClientBootstrap;
+import io.netty.bootstrap.Bootstrap;
 import io.netty.buffer.ByteBuf;
 import io.netty.channel.Channel;
 import io.netty.channel.ChannelFuture;
@@ -43,7 +43,7 @@ public class HexDumpProxyFrontendHandler extends ChannelInboundByteHandlerAdapte
         final Channel inboundChannel = ctx.channel();
 
         // Start the connection attempt.
-        ClientBootstrap b = new ClientBootstrap();
+        Bootstrap b = new Bootstrap();
         b.group(inboundChannel.eventLoop())
          .channel(NioSocketChannel.class)
          .remoteAddress(remoteHost, remotePort)
