+++ example/src/main/java/io/netty/example/sctp/SctpEchoClient.java
@@ -15,8 +15,8 @@
  */
 package io.netty.example.sctp;
 
+import io.netty.bootstrap.AbstractBootstrap;
 import io.netty.bootstrap.Bootstrap;
-import io.netty.bootstrap.ClientBootstrap;
 import io.netty.channel.ChannelFuture;
 import io.netty.channel.ChannelInitializer;
 import io.netty.channel.ChannelOption;
@@ -52,7 +52,7 @@ public class SctpEchoClient {
 
     public void run() throws Exception {
         // Configure the client.
-        ClientBootstrap b = new ClientBootstrap();
+        Bootstrap b = new Bootstrap();
         try {
             b.group(new NioEventLoopGroup())
              .channel(NioSctpChannel.class)
