+++ example/src/main/java/io/netty/example/uptime/UptimeClient.java
@@ -15,7 +15,7 @@
  */
 package io.netty.example.uptime;
 
-import io.netty.bootstrap.ClientBootstrap;
+import io.netty.bootstrap.Bootstrap;
 import io.netty.channel.ChannelInitializer;
 import io.netty.channel.EventLoopGroup;
 import io.netty.channel.socket.SocketChannel;
@@ -50,14 +50,14 @@ public class UptimeClient {
     }
 
     public void run() {
-        configureBootstrap(new ClientBootstrap()).connect();
+        configureBootstrap(new Bootstrap()).connect();
     }
 
-    private ClientBootstrap configureBootstrap(ClientBootstrap b) {
+    private Bootstrap configureBootstrap(Bootstrap b) {
         return configureBootstrap(b, new NioEventLoopGroup());
     }
 
-    ClientBootstrap configureBootstrap(ClientBootstrap b, EventLoopGroup g) {
+    Bootstrap configureBootstrap(Bootstrap b, EventLoopGroup g) {
         b.group(g)
          .channel(NioSocketChannel.class)
          .remoteAddress(host, port)
