+++ testsuite/src/test/java/io/netty/testsuite/transport/socket/DatagramMulticastTest.java
@@ -16,7 +16,7 @@
 package io.netty.testsuite.transport.socket;
 
 import static org.junit.Assert.*;
-import io.netty.bootstrap.Bootstrap;
+import io.netty.bootstrap.AbstractBootstrap;
 import io.netty.buffer.Unpooled;
 import io.netty.channel.Channel;
 import io.netty.channel.ChannelHandlerContext;
@@ -40,7 +40,7 @@ public class DatagramMulticastTest extends AbstractDatagramTest {
         run();
     }
 
-    public void testMulticast(Bootstrap sb, Bootstrap cb) throws Throwable {
+    public void testMulticast(AbstractBootstrap sb, AbstractBootstrap cb) throws Throwable {
         MulticastTestHandler mhandler = new MulticastTestHandler();
 
         sb.handler(new ChannelInboundMessageHandlerAdapter<DatagramPacket>() {
